Add files via upload
It took about 7 hours to get this to work. making the lines gray while keeping the white dots wound up being a challenging process from which I encountered several bugs. eventually i realized i could do it all at once, multiplying the matrix by 0.5 and making a double for loop to re-add the white dots and removing the end row and column to make it look nicer. then i implemented it with Psychtoolbox